backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
groq
playwright
webdriver-manager
opencv-python
numpy
ImageHash
python-whois
pydantic
sqlalchemy
supabase
python-multipart
Pillow
python-Levenshtein
python-magic
